# api/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Mobile AI Coding Monitor \u0026 Controller Agent API.\nEnables remote monitoring and control of Claude Code CLI sessions from mobile devices.",
        "title": "cdev API",
        "contact": {
            "name": "Brian Ly",
            "url": "https://github.com/brianly1003/cdev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8766",
    "basePath": "/",
    "paths": {
...
        "/api/repository/files/list": {
            "get": {
                "description": "Returns a paginated list of files in a directory with filtering and sorting options",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "List files in directory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Directory path (empty for root)",
                        "name": "directory",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include subdirectories",
                        "name": "recursive",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "name",
                        "description": "Sort by: name, size, modified",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "asc",
                        "description": "Sort order: asc, desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated file extensions to filter",
                        "name": "extensions",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.FileList"
                        }
                    },
                    "503": {
                        "description": "Repository indexer not available",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/repository/files/tree": {
            "get": {
                "description": "Returns a hierarchical tree structure of the repository",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Get directory tree",
                "parameters": [
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "Root path",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 2,
                        "description": "Maximum depth",
                        "name": "depth",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.DirectoryTree"
                        }
                    },
                    "503": {
                        "description": "Repository indexer not available",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/repository/index/rebuild": {
            "post": {
                "description": "Triggers a full re-index of the repository",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Rebuild repository index",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Repository indexer not available",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/repository/index/status": {
            "get": {
                "description": "Returns the current status of the repository index",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Get repository index status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.IndexStatus"
                        }
                    },
                    "503": {
                        "description": "Repository indexer not available",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/repository/search": {
            "get": {
                "description": "Search for files using fuzzy, exact, prefix, or extension matching",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Search files in repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "fuzzy",
                        "description": "Search mode: fuzzy, exact, prefix, extension",
                        "name": "mode",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Maximum results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated file extensions to filter",
                        "name": "extensions",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Exclude binary files",
                        "name": "exclude_binaries",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Only include git-tracked files",
                        "name": "git_tracked_only",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.SearchResult"
                        }
                    },
                    "400": {
                        "description": "Missing query parameter",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Repository indexer not available",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/repository/stats": {
            "get": {
                "description": "Returns aggregate statistics about the repository",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Get repository statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.RepositoryStats"
                        }
                    },
                    "503": {
                        "description": "Repository indexer not available",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                },
                "path": {
                    "type": "string",
                    "example": "src/main.ts"
                }
            }
        },
...
        "repository.DirectoryInfo": {
            "type": "object",
            "properties": {
                "file_count": {
                    "type": "integer"
                },
                "last_modified": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "total_size_bytes": {
                    "type": "integer"
                }
            }
        },
        "repository.DirectoryTree": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.DirectoryTree"
                    }
                },
                "extension": {
                    "type": "string"
                },
                "file_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "total_size_bytes": {
                    "type": "integer"
                },
                "type": {
                    "description": "\"file\" or \"directory\"",
                    "type": "string"
                }
            }
        },
        "repository.FileInfo": {
            "type": "object",
            "properties": {
                "directory": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "git_ignored": {
                    "type": "boolean"
                },
                "git_tracked": {
                    "type": "boolean"
                },
                "indexed_at": {
                    "type": "string"
                },
                "is_binary": {
                    "type": "boolean"
                },
                "is_sensitive": {
                    "type": "boolean"
                },
                "is_symlink": {
                    "type": "boolean"
                },
                "line_count": {
                    "type": "integer"
                },
                "match_score": {
                    "description": "For search results",
                    "type": "number"
                },
                "modified_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                }
            }
        },
        "repository.FileList": {
            "type": "object",
            "properties": {
                "directories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.DirectoryInfo"
                    }
                },
                "directory": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.FileInfo"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/repository.PaginationInfo"
                },
                "total_directories": {
                    "type": "integer"
                },
                "total_files": {
                    "type": "integer"
                }
            }
        },
        "repository.IndexStatus": {
            "type": "object",
            "properties": {
                "database_size_bytes": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                },
                "indexed_files": {
                    "type": "integer"
                },
                "is_git_repo": {
                    "type": "boolean"
                },
                "last_full_scan": {
                    "type": "string"
                },
                "last_update": {
                    "type": "string"
                },
                "status": {
                    "description": "ready, indexing, error",
                    "type": "string"
                },
                "total_files": {
                    "type": "integer"
                },
                "total_size_bytes": {
                    "type": "integer"
                }
            }
        },
        "repository.PaginationInfo": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
        "repository.RepositoryStats": {
            "type": "object",
            "properties": {
                "binary_files": {
                    "type": "integer"
                },
                "files_by_extension": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "git_ignored_files": {
                    "type": "integer"
                },
                "git_tracked_files": {
                    "type": "integer"
                },
                "largest_files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.FileInfo"
                    }
                },
                "sensitive_files": {
                    "type": "integer"
                },
                "total_directories": {
                    "type": "integer"
                },
                "total_files": {
                    "type": "integer"
                },
                "total_size_bytes": {
                    "type": "integer"
                }
            }
        },
        "repository.SearchMode": {
            "type": "string",
            "enum": [
                "fuzzy",
                "exact",
                "prefix",
                "extension"
            ],
            "x-enum-varnames": [
                "SearchModeFuzzy",
                "SearchModeExact",
                "SearchModePrefix",
                "SearchModeExtension"
            ]
        },
        "repository.SearchResult": {
            "type": "object",
            "properties": {
                "elapsed_ms": {
                    "type": "integer"
                },
                "mode": {
                    "$ref": "#/definitions/repository.SearchMode"
                },
                "query": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.FileInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
...
}